requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
mysql-connector-python==8.0.33
pillow==11.1.0
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.12.2
gunicorn==23.0.0
dj-database-url==2.1.0
mysqlclient>=2.2.0
whitenoise==6.6.0
django_filters==2.0.0